requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
attrs==24.3.0
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cn2an==0.5.23
crcmod==1.7
cryptography==44.0.0
datasets==3.0.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
editdistance==0.8.1
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.6.1
funasr==1.2.3
google-genai==1.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
jiwer==3.0.5
jmespath==0.10.0
joblib==1.4.2
jsonlines==4.0.0
kaldiio==2.18.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
modelscope==1.22.3
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.60.0
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pip==24.2
platformdirs==4.3.6
pooch==1.8.2
proces==0.1.7
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytorch-wpe==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.0
soxr==0.5.0.post1
sympy==1.13.3
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torch-complex==0.4.4
tqdm==4.67.1
transformers==4.44.1
triton==3.0.0
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
wheel==0.44.0
xxhash==3.5.0
yarl==1.18.3
zhon==2.1.1
